Squashed commit of the following:
commit 3611f92c04dfb9f65b2ec566887142a5c5e809a8
Author: liam-koh <144200505+liam-koh@users.noreply.github.com>
Date:   Wed Feb 26 12:00:31 2025 +0900

    Rft: 기존 bingnewsfetch 코드 개선코드로 교체
