+++ src/components/organisms/search/NewsSearchGridList/index.tsx
@@ -1,17 +1,16 @@
 'use client';
 
 import NewsGridList from '@/components/molecules/NewsGridList';
-import useFetchNewsList from '@/hooks/useFetchNewsList';
 import React, { useMemo } from 'react';
 import useBingNewsFetch from '@/queries/useBingNewsFetch';
 import NewsCard from '../NewsCard';
 
 export default function NewsSearchGridList() {
-  // const { data } = useFetchNewsList.query();
-  const { filterQueries } = useFetchNewsList.state();
+  const { filterQueries } = useBingNewsFetch.state();
+  // TODO: query 관련코드 제거
   const query = filterQueries.keyword;
   const isQueryEmpty = useMemo(() => query === '', [query]);
-  const queryStates = useBingNewsFetch({
+  const queryStates = useBingNewsFetch.query({
     query,
     enabled: !isQueryEmpty,
     maxPage: 3,
